+++ scripts/generate-app-config/generateNameConfig.js
@@ -8,6 +8,8 @@ const generateNameConfig = (appConfig) => {
   const imports = new ImportAggregator();
   const config = {
     title: startCase(appConfig.title),
+    ...appConfig.headline && { headline: appConfig.headline },
+    ...appConfig.subline && { subline: appConfig.subline },
   };
 
   return { config, imports };