+++ CHANGELOG.md
@@ -3,6 +3,20 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* Updated `commander` to `v3.0.1`
+* Terra-dev-site now looks to the webpack resolve extensions to add to the files pulled in when the site is generated.
+
+6.2.0 - (August 30, 2019)
+----------
+### Changed
+* Updated Object.assign syntax to Object spread syntax
+* Updated `keycode-js` to `V2.0.1`
+* auto focus the search field
+* Provide a webpack plugin for consumption vs a webpack config.
+
+### Fixed
+* locale override was not being applied.
 
 6.1.1 - (August 13, 2019)
 ----------